+++ lib/paramsync/cli/check_command.rb
@@ -1,11 +1,11 @@
 # This software is public domain. No rights are reserved. See LICENSE for more information.
 
-class Constancy
+class Paramsync
   class CLI
     class CheckCommand
       class << self
         def run(args)
-          Constancy::CLI.configure
+          Paramsync::CLI.configure
 
           mode = if args.include?("--pull")
                    :pull
@@ -13,7 +13,7 @@ class Constancy
                    :push
                  end
 
-          Constancy.config.sync_targets.each do |target|
+          Paramsync.config.sync_targets.each do |target|
             diff = target.diff(mode)
             diff.print_report
             if not diff.any_changes?