+++ lib/paramsync/cli/targets_command.rb
@@ -1,13 +1,13 @@
 # This software is public domain. No rights are reserved. See LICENSE for more information.
 
-class Constancy
+class Paramsync
   class CLI
     class TargetsCommand
       class << self
         def run
-          Constancy::CLI.configure(call_external_apis: false)
+          Paramsync::CLI.configure(call_external_apis: false)
 
-          Constancy.config.sync_targets.each do |target|
+          Paramsync.config.sync_targets.each do |target|
             if target.name
               puts target.name
             else