+++ spec/lib/paramsync/config_spec.rb
@@ -2,20 +2,20 @@
 
 require 'spec_helper'
 
-RSpec.describe Constancy::Config do
+RSpec.describe Paramsync::Config do
   describe '#only_valid_config_keys!' do
     context 'taking lists of valid configuration keys' do
       let(:keylists) {
         [
-          %w( constancy vault consul sync ),
-          %w( constancy vault vault.main vault.dev sync ),
+          %w( paramsync vault consul sync ),
+          %w( paramsync vault vault.main vault.dev sync ),
           %w( vault.us-east-1 vault.b220300 ),
         ]
       }
 
       it 'accepts all valid key lists' do
         keylists.each do |keylist|
-          expect(Constancy::Config.only_valid_config_keys!(keylist)).to be true
+          expect(Paramsync::Config.only_valid_config_keys!(keylist)).to be true
         end
       end
     end
@@ -23,8 +23,8 @@ RSpec.describe Constancy::Config do
     context 'taking lists containing invalid configuration keys' do
       let(:keylists) {
         [
-          %w( constancy vault consul sync.banana ),
-          %w( constancy vault vault.main vault._dev sync ),
+          %w( paramsync vault consul sync.banana ),
+          %w( paramsync vault vault.main vault._dev sync ),
           %w( vault.us-east-1 vault.220300 ),
           %w( volt console sink constancee ),
         ]
@@ -33,8 +33,8 @@ RSpec.describe Constancy::Config do
       it 'raises an exception for invalid key lists' do
         keylists.each do |keylist|
           expect {
-            Constancy::Config.only_valid_config_keys!(keylist)
-          }.to raise_exception(Constancy::ConfigFileInvalid, /is not a valid configuration key/)
+            Paramsync::Config.only_valid_config_keys!(keylist)
+          }.to raise_exception(Paramsync::ConfigFileInvalid, /is not a valid configuration key/)
         end
       end
     end